+++ ui/src/network/forms/index.js
@@ -1,0 +1,4 @@
+
+export {NetworkForm} from './network';
+export {NetworkProxiesForm} from './proxies';
+export {NetworkServicesForm} from './services';
